+++ styles/global.css
@@ -1,0 +1,5 @@
+body {
+  margin: 0;
+  padding: 0;
+  background: url(/background.jpg) no-repeat center center fixed; 
+}
